snackbar as composed reactive
